pyproject.toml
[tool.poetry]
name = "rtxscraper"
version = "0.3.3"
description = ""
authors = ["nlazcani"]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^3.141.0"
webdriver-manager = "^3.4.2"
python-telegram-bot = "^13.7"

[tool.poetry.dev-dependencies]
black = {version = "^21.7b0", allow-prereleases = true}
pycodestyle = "^2.7.0"
pylint = "^2.9.6"
pylama = "^7.7.1"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
